requirements.txt
tensorflow==2.3.0
Pillow==7.2.0
numpy==1.18.5
gensim==3.8.3
<<<<<<< HEAD
=======
streamlit==0.67.0
>>>>>>> 47cdcde05df520c338fb899663494fb4c82ec0ce
matplotlib>=3.3.2
